pyproject.toml
[project]
name = "raycast-dspy-backend"
version = "0.1.0"
description = "DSPy backend for Raycast prompt improver"
readme = "docs/backend/README.md"
requires-python = ">=3.10"
dependencies = [
  "dspy-ai>=3.0.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "litellm>=1.0.0",
  "python-dotenv>=1.0.0",
  "requests>=2.32.0",
]

[dependency-groups]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
]
